backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
alembic
PyGithub>=1.55
pytest
pytest-mock
httpx
psutil
psycopg2-binary
requests
google-generativeai
apscheduler
python-jose[cryptography]
passlib[bcrypt]
python-multipart
ruff
mypy
types-requests
hoverfly
rich
